[bugfix] pin MeanMetric all_reduce tensor to float64 to avoid cross-rank dtype mismatch
MeanMetric.compute() built the reduction tensor with an inferred dtype:
`torch.tensor([self.state, self.count], device=...)`. When sequence-level
`compute_acc()` accumulates `np.all(...)` (a numpy.bool_), it promotes
self.state from python float to numpy.float64 on those ranks, so torch infers
float64 there but float32 on ranks that stayed python float. The ranks then
enter `dist.all_reduce()` with mismatched dtypes and the collective hangs
(#9550).

Pin the tensor to float64 so every rank reduces with the same dtype regardless
of self.state's runtime type; float64 also keeps count exact for large totals.

+++ swift/metrics/utils.py
@@ -101,7 +101,7 @@ class MeanMetric(Metric):
 
     def compute(self):
         if dist.is_initialized():
-            tensor = torch.tensor([self.state, self.count], device=self.device)
+            tensor = torch.tensor([self.state, self.count], dtype=torch.float64, device=self.device)
             dist.all_reduce(tensor, op=dist.ReduceOp.SUM, group=self.group)
             self.state, self.count = tensor[0].item(), int(tensor[1].item())
         if self.count == 0:
